+++ lib/flutter_artist_core.dart
@@ -34,7 +34,7 @@ part 'src/_log/_log_entry.dart';
 part 'src/_log/_log_summary.dart';
 part 'src/_part/_api_result.dart';
 part 'src/_part/_converters.dart';
-part 'src/_part/_fa_data_converters.dart';
+part 'src/_part/_fa_item_converters.dart';
 part 'src/_part/_fa_json_converters.dart';
 part 'src/_part/_identifiable.dart';
 part 'src/_part/_list_data.dart';